chore(core/vm): clean up initial process
